tests/krate/publish: Extract `similar_names` module
